use reduce utils
